Shiny Application
This application is about about the world distribution of telephone by region 
from 1951 to 1961.
To use the application you have to click on the button to the left. if it cannot show ,the data on the left is the number of telephone  distributed as you can see on the server.r.
To see the different countries and their variation with the others countries
on the right hand side you will also  see how the usersinterface and the server  vary according to each country. The data on the bottom  on the x-axis is the number of years during the decade of distribution.